pyproject.toml
[project]
name = "wukong-voice-agent"
version = "1.0.0"
description = "WuKong AI Voice Tutor - An AI-powered voice assistant for explaining DeepSeek MLA"
requires-python = ">=3.10,<3.14"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.3",
    "python-dotenv",
]

[project.scripts]
agent = "agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
